requirements.txt
python-dotenv
numpy
pandas
cohere
scikit-learn